requirements.txt
torchvision==0.9.0
torchaudio==0.8.0
matplotlib
easydict
librosa==0.8.1
essentia
chumpy
transforms3d
pyarrow==0.14.1
lmdb==0.96
numpy
configargparse
tensorboard
setuptools==59.5.0
loguru
fasttext
textgrid
Ipython
Cython
nnAudio
opencv-python
pandas
parso
Pillow
pyrsistent
PyWavelets
pyzmq
scikit-image
scikit-learn
scipy
six
terminado
testpath
timm
yarg
yarl
tqdm
ffmpeg-normalize
pyworld
python_speech_features
sox
fairseq
python-Levenshtein
seaborn
sentence-transformers
